Move classes to adapters package
