+++ lib/flutter_formx.dart
@@ -1,8 +1,8 @@
-export 'src/form/bloc/formx_cubit.dart';
+export 'src/form/adapters/bloc/formx_cubit.dart';
+export 'src/form/adapters/mobx/formx_mobx.dart';
 export 'src/form/core/formx.dart';
 export 'src/form/core/formx_field.dart';
 export 'src/form/core/formx_state.dart';
-export 'src/form/mobx/formx_mobx.dart';
 export 'src/validator/required_field_validator.dart';
 export 'src/validator/validator.dart';
 export 'src/validator/validator_result.dart';
