+++ lib/src/form/adapters/mobx/formx_mobx.dart
@@ -1,6 +1,6 @@
+import 'package:flutter_formx/src/form/adapters/formx_adapter.dart';
 import 'package:flutter_formx/src/form/core/formx.dart';
 import 'package:flutter_formx/src/form/core/formx_field.dart';
-import 'package:flutter_formx/src/form/core/formx_adapter.dart';
 import 'package:flutter_formx/src/form/core/formx_state.dart';
 import 'package:mobx/mobx.dart';
 
@@ -26,10 +26,9 @@ mixin FormXMobX<T> implements FormXAdapter<T> {
   }
 
   @override
-  Future<void> updateAndValidateField(newValue, type,
-      {bool softValidation = false}) async {
-    final validatedField = await _formX.value
-        .updateAndValidateField(newValue, type, softValidation: softValidation);
+  Future<void> updateAndValidateField(newValue, type, {bool softValidation = false}) async {
+    final validatedField =
+        await _formX.value.updateAndValidateField(newValue, type, softValidation: softValidation);
     Action(() {
       _formX.value = validatedField;
     })();
@@ -44,8 +43,7 @@ mixin FormXMobX<T> implements FormXAdapter<T> {
 
   @override
   Future<bool> validateForm({bool softValidation = false}) async {
-    final validatedForm =
-        await _formX.value.validateForm(softValidation: softValidation);
+    final validatedForm = await _formX.value.validateForm(softValidation: softValidation);
     Action(() {
       _formX.value = validatedForm;
     })();
